Add files via upload
dataset:
Columns
File: Unit_1_Project_Dataset (1).csv

Column name	Description
host_name	Name of the host of the listing. (String)
host_since_year	Year the host joined AirBnb. (Integer)
host_since_anniversary	Anniversary of the host joining AirBnb. (Integer)
neighbourhood_cleansed	The neighbourhood the listing is located in. (String)
city	The city the listing is located in. (String)
state	The state the listing is located in. (String)
zipcode	The zipcode of the listing. (Integer)
country	The country the listing is located in. (String)
latitude	The latitude of the listing. (Float)
longitude	The longitude of the listing. (Float)
property_type	The type of property the listing is. (String)
room_type	The type of room the listing is. (String)
accommodates	The number of people the listing can accommodate. (Integer)
bathrooms	The number of bathrooms the listing has. (Integer)
bedrooms	The number of bedrooms the listing has. (Integer)
beds	The number of beds the listing has. (Integer)
bed_type	The type of bed the listing has. (String)
price	The price of the listing. (Float)
guests_included	The number of guests included in the price. (Integer)
extra_people	The additional cost for extra people. (Float)
minimum_nights	The minimum number of nights a guest must stay. (Integer)
host_response_time	The time it takes for the host to respond to a request. (String)
host_response_rate	The rate at which the host responds to requests. (Float)
number_of_reviews	The number of reviews the listing has. (Integer)
review_scores_rating	The overall rating of the listing. (Float)
review_scores_accuracy	The accuracy rating of the listing. (Float)